requirements.txt
argparse==1.2.1
numpy==1.10.4
scipy==0.16.1
nd2reader==2.0.2
biopython==1.66
cycler==0.9.0
dill==0.2.4
matplotlib==1.5.1
scikit-learn==0.17
ppft==1.6.4.5
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
tifffile==0.7.0
wsgiref==0.1.2
xmltodict==0.9.2
pysam==0.8.4
pyparsing==2.0.7
pyre==0.8-pathos
pathos==0.2a1.dev0
multiprocess==0.70.3
pox==0.2.2
pp==1.5.7-pathos
